[stats] handle all stats as int64

+++ backends/abstract.go
@@ -52,6 +52,6 @@ type Abstract interface {
 	DeleteGroup(groupgid string) *Error
 	Groups() ([]Group, *Error)
 	GroupUsers(groupgid string) ([]User, *Error)
-	Stats() (stats map[string]int, err *Error)
+	Stats() (stats map[string]int64, err *Error)
 	Close()
 }
